Generate static forwarders for object members in companion interface
We used to disable generation of static forwarders when a object had a
trait as a companion, as one could not add methods with bodies to an
interface in JVM 6.

The JVM lifted this restriction to support default methods in interfaces,
so we can lift the restriction on static forwarders, too.

Fixes https://github.com/scala/scala-dev/issues/59