Generate static forwarders for object members in companion interface
We used to disable generation of static forwarders when a object had a
trait as a companion, as one could not add methods with bodies to an
interface in JVM 6.

The JVM lifted this restriction to support default methods in interfaces,
so we can lift the restriction on static forwarders, too.

I've only commited the test using the default backend (GenBCode), but
I've also made the change and manaully verified the test works under
the soon-to-be-removed GenASM.

```
% qscalac -Ybackend:GenASM test/files/run/trait-static-forwarder/forwarders.scala && javac -d . -classpath . test/files/run/trait-static-forwarder/Test.java && qscala Test
./T.class: warning: Cannot find annotation method 'bytes()' in type 'ScalaSignature': class file for scala.reflect.ScalaSignature not found
1 warning
42
```

Fixes https://github.com/scala/scala-dev/issues/59

I had to disable use of the lambda deserialization refletion cache
when in interfaces, as this change revealed a latent design issue
with our lambda deserialization approach. Disabling the cache isn't
a correctness problem, but hampers performance.