SI-9527 Fix NPE in ambiguous implicit error generation
In #4673, an annotation was added to allow customization of the
compiler error for ambigous implicits. This change had incorrect
assumptions about the shape of trees that would be generated during
implicit search, and failed to account for the results of eta expanded
methods when searching for a function type.

This commit:

  - Uses the symbol from `ImpilcitInfo`, rather than calling
    `Tree#symbol` which is fraught with the danger of returning a
    null symbol for something other than an application.
  - Adds a test for customized messages for a polymorphic, eta
    expanded method, and generalizes `treeTypeArgs` to handle the
    implicit tree of shape `{ (x: X) => f[A](x)}`.