API only tracks declared type of self variable
The only aspect of the self variable that's relevant for
incremental compilation is its explicitly declared type,
and only when it's different from the type of the class that declares it.

Technically, any self type that's a super type of the class could be ignored,
as it cannot affect external use (instantiation/subclassing) of the class.

Rewritten from sbt/zinc@cc0e66ee6d5b55a5c23843da5d44daf3e1076881