Consider signatures of method before and after erasure in ExtractAPI
The signatures of methods that have value classes as arguments or return
type change during the erasure phase. Because we only registered
signatures before the erasure, we missed some API changes when a class
was changed to a value class (or a value class changed to a class).

This commit fixes this problem by recording the signatures of method
before and after erasure.

Fixes sbt/sbt#1171

Rewritten from sbt/zinc@1980cd32303d2879fd08bca18893aea8a44dc5aa