SI-9542 Fix regression in value classes with enclosing refs
In SI-9473 / e2653736, I changed `mkAttributedRef` to avoid prefixing
references to statically owned symbols with a `ThisType`.

Turns out there was one place that depended on the old behaviour.

```
object Outer {
  trait T
  class C(val value: Any) extends AnyVal {
	def foo(t: T) = expr
  }
}
```

Is translated to:

```
object Outer {
  trait T
  class C(val value: Any) extends AnyVal {
	def foo(t: T) = Outer.this.C.foo$extension(C.this, t)
  }
  object C {
    def foo$extension($this: C, t: T) = expr
  }
}
```

After the change, the forwarder was instead:

```
	def foo(t: T) = Outer.C.foo$extension(C.this, t)
```

Note: this change is not actually necessary after the following
commit that makes subtyping unify the different module class
reference, that alone is enough to make the enclosed test
pass.