DRY-er trees in pattern matcher code gen.
Rather than building a cascade of if/elses, push additional conditions
into a conjunction in the condition of a single if/else. This is
possible when emitting conditions for the list of arguments of a
pattern.

Here's an example of the improvement to post-pattern matcher trees:

  https://gist.github.com/retronym/0d8f7126157061d72b81

While we could try to rely on the optimizer to coalesce the repeated
else clauses, it seems wasteful to emit the code in that way in
the first place.