wip: Constant Folding preserves original tree
Introduce `Const` extractor to unify scenarios in which tree is a constant.
The constant may be found directly in the tree (it was a constant originally),
or its type (due to constant folding).

Instead of typer, uncurry now replaces trees of ConstantType with the
corresponding literal. We do this in uncurry and not erasure, because
the intervening explicitouter and tailcalls should not see
references to members holding constants, as they will be folded to
the constant in byte code.

Explicit outer would unnecessarily make private final vals not-private
if it saw references to them (instead of the constant).