Improved EOL handling for scripts generated by the sbt build
We now normalize EOLs in `ScalaTool` to LF when reading the templates,
and write them as CRLF or LF depending on the target platform. This
allows the script templates to be stored locally with either LF or CRLF
line endings (which can both happen, depending on the platform and git
configuration).