SI-8044 Allow binding backquoted varid in patterns
Previously, a varid could not be backquoted, so that it was not
possible to introduce variables with names such as `type` in a
match expression.

This commit allows backquoted varids in `case x @ _` and
`case x: Int`. In neither position is a stable id accepted,
that is, an id with leading uppercase.

Therefore, this commit merely relaxes the backquoted varid to
be taken as a normal varid in these contexts.