Improve simplifyJumps Improve simplifyJumps to rewrite
  IFEQ L4
  L5
  GOTO L6

to

  IFNE L6
  L5

This rewrite is only correct if L5 is not the target of any jump
instruction (otherwise, removing the GOTO would change semantics).
Previously we did not do the rewrite if there was any label between
the conditional jump and the goto (like L5). Now we track which labels
are jump targets.