Enable IntellIJ smarts while editing the build
Even though our build can't be imported into IntelliJ
automatically, we don't need to live in the dark ages
of red squigglies when editing build.sbt.

This commit hand-crafts the module defintion for the
project, which has SBT and its dependendencies on the
classpath.

A screenshot of the code assist is worth the thousand
expletives I uttered while writing the pictured code
without this facility:

  https://www.dropbox.com/s/6mxv0iwxkhvnor7/Screenshot%202016-02-07%2022.09.12.png?dl=0

To download the sources for the SBT JARs that are referenced
herein, run

    sbt> reload plugins
    sbt> updateClassifiers